+++ config/app.php
@@ -18,7 +18,7 @@ return [
 
 	'product_url' => 'https://github.com/mrcage/ohf-community',
 
-    'version' => 'v1.22.3',
+    'version' => 'v1.23.0',
     
 
     /*